Switch to bfloat16.

+++ tools/gen_search_space.py
@@ -49,7 +49,7 @@ class GemmNormalSchema:
     impl_header = "gemm_normal/gemm_v2_impl.h"
     
     def get_meta_space(self, w):
-        data_type = [('FP16', 'FP16', 'FP16', 'FP32')] # a,b,cd,acc
+        data_type = [('BF16', 'BF16', 'BF16', 'FP32')] # a,b,cd,acc
         layout = ['RCR'] # , 'RRR'
         arch = ['Sm80'] # , 'Sm89'
 
@@ -103,7 +103,7 @@ class GemmNormalSimtSchema:
     impl_header = "gemm_normal/gemm_v2_simt_impl.h"
     
     def get_meta_space(self, w):
-        data_type = [('FP16', 'FP16', 'FP16', 'FP32')] # a,b,cd,acc
+        data_type = [('BF16', 'BF16', 'BF16', 'FP32')] # a,b,cd,acc
         layout = ['RCR'] # , 'RRR'
         arch = ['Sm80'] # , 'Sm89'
